Time: 31 ms (10.66%), Space: 5.9 MB (25.17%) - LeetHub

+++ 69-sqrtx/69-sqrtx.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
-    int mySqrt(int x) {
-     long i=0;
+    int mySqrt(long long x) {
+     long long i=0;
         while(i*i <= x){
             i++;
         }
